SpotLight Added

+++ Renderer/DirectionalLight.hpp
@@ -7,6 +7,6 @@ struct DirectionalLight
 	Vector3 color;
 	float intensity;
 
-	DirectionalLight(Vector3 color, int intensity) : color(color), intensity(intensity) {};
+	DirectionalLight(Vector3 color, float intensity) : color(color), intensity(intensity) {};
 	DirectionalLight() : color(1.0f, 1.0f, 1.0f), intensity(1.0f) {};
 };
